Moved ghostbusters file under movies directory and class is now scoped to a movies module.
